requirements.txt
boto3==1.26.2
psycopg[binary]~=3.1.17
requests==2.31.0
bs4~=0.0.2

lxml~=5.1.0
botocore~=1.34.34
Pillow~=10.2.0
beautifulsoup4~=4.12.3
selenium~=4.17.2

pytest~=8.0.0
